pyproject.toml
[project]
name = "blocker-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest-asyncio>=1.3.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
